+++ tests/PyroTests/test_core.py
@@ -274,53 +274,6 @@ class CoreTests(unittest.TestCase):
         self.assertFalse(Pyro4.core.URI.isUnixsockLocation("./x:name"))
         self.assertFalse(Pyro4.core.URI.isUnixsockLocation("foobar"))
 
-    def testMsgFactory(self):
-        import hashlib, hmac
-        def pyrohmac(data):
-            return hmac.new(Pyro4.config.HMAC_KEY, data, digestmod=hashlib.sha1).digest()
-        MF=Pyro4.core.MessageFactory
-        MF.createMessage(99, None, 0,0) # doesn't check msg type here
-        self.assertRaises(Pyro4.errors.ProtocolError, MF.parseMessageHeader, "FOOBAR")
-        hdr=MF.createMessage(MF.MSG_CONNECT, b"hello",0,0)[:-5]
-        msgType,flags,seq,dataLen,datahmac=MF.parseMessageHeader(hdr)
-        self.assertEqual(MF.MSG_CONNECT, msgType)
-        self.assertEqual(MF.FLAGS_HMAC, flags)
-        self.assertEqual(5, dataLen)
-        self.assertEqual(pyrohmac(b"hello"), datahmac)
-        hdr=MF.createMessage(MF.MSG_RESULT, None,0,0)
-        msgType,flags,seq,dataLen,datahmac=MF.parseMessageHeader(hdr)
-        self.assertEqual(MF.MSG_RESULT, msgType)
-        self.assertEqual(MF.FLAGS_HMAC, flags)
-        self.assertEqual(0, dataLen)
-        hdr=MF.createMessage(MF.MSG_RESULT, b"hello", 42, 0)[:-5]
-        msgType,flags,seq,dataLen,datahmac=MF.parseMessageHeader(hdr)
-        self.assertEqual(MF.MSG_RESULT, msgType)
-        self.assertEqual(42, flags)
-        self.assertEqual(5, dataLen)
-        msg=MF.createMessage(255,None,0,255)
-        self.assertEqual(38,len(msg))
-        msg=MF.createMessage(1,None,0,255)
-        self.assertEqual(38,len(msg))
-        msg=MF.createMessage(1,None,flags=253,seq=254)
-        self.assertEqual(38,len(msg))
-        # compression is a job of the code supplying the data, so the messagefactory should leave it untouched
-        data=b"x"*1000
-        msg=MF.createMessage(MF.MSG_INVOKE, data, 0,0)
-        msg2=MF.createMessage(MF.MSG_INVOKE, data, MF.FLAGS_COMPRESSED,0)
-        self.assertEqual(len(msg),len(msg2))
-
-    def testMsgFactoryProtocolVersion(self):
-        version=Pyro4.constants.PROTOCOL_VERSION
-        Pyro4.constants.PROTOCOL_VERSION=0     # fake invalid protocol version number
-        msg=Pyro4.core.MessageFactory.createMessage(Pyro4.core.MessageFactory.MSG_RESULT, b"result", 0, 1)
-        try:
-            Pyro4.core.MessageFactory.parseMessageHeader(msg)
-            self.fail("expected protocolerror")
-        except Pyro4.errors.ProtocolError:
-            pass
-        finally:
-            Pyro4.constants.PROTOCOL_VERSION=version
-        
     def testProxyOffline(self):
         # only offline stuff here.
         # online stuff needs a running daemon, so we do that in another test, to keep this one simple